Вывод эмоций на главный фрагмент, +фикс Dao и ViewModel

+++ app/src/main/java/com/example/moodnote/data/MoodRepository.kt
@@ -24,7 +24,7 @@ class MoodRepository @Inject constructor(
         noteDao.insertOrReplace(note)
     }
 
-    fun getNote(id: Int): Note {
+    fun getNote(id: Long): Note {
         return noteDao.getNote(id)
     }
